\name{bdt}
\alias{bdt}
\docType{package}
\title{Default object for RcppBDT Boost Date_Time binding}
\description{
  The \verb{bdt} variable is a default instance of the reference class
  created by means of Rcpp modules.

  New instances can be created using either the default constructor
  (without argument) or the constructor using year, month, date arguments.
}
\details{
  Please consult the Boost documentation for (copious) details on the
  Date_Time library. See the Rcpp-modules vignette for details on Rcpp
  modules. 
}
\author{Dirk Eddelbuettel \email{edd@debian.org}}
\references{Boost Date_Time: \url{http://www.boost.org/doc/html/date_time.html}}
\keyword{package}
