\name{bdtEnv}
\alias{bdtEnv}
\docType{package}
\title{Bindings for Boost Date_Time}
\description{
  The \verb{bdtEnv} environment contains the reference class instance
  created by means of Rcpp modules, as well as an object instance.
}
\details{
  Please consult the Boost documentation for (copious) details on the
  Date_Time library. See the Rcpp-modules vignette for details on Rcpp
  modules. 
}
\author{Dirk Eddelbuettel \email{edd@debian.org}}
\references{Boost Date_Time: \url{http://www.boost.org/doc/html/date_time.html}}
\keyword{package}
