\name{RcppBDT-package}
\alias{RcppBDT-package}
\alias{RcppBDT}
\docType{package}
\title{
  Bindings for Boost Date_Time
}
\description{
  This package provides R with access to Boost Date_Time
  functonality by using Rcpp modules. Currently only Date functionality 
  is covered.  
}
\details{
\tabular{ll}{
Package: \tab RcppBDT\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab $Date$\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
  Dirk Eddelbuettel <edd@debian.org>
  Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\references{
  Boost Date_Time: \url{http://www.boost.org/doc/html/date_time.html}
}
\keyword{package}

