% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parsePOSIXt}
\alias{parsePOSIXt}
\title{Parse POSIXct objects from character variables}
\usage{
parsePOSIXt(sv)
}
\arguments{
\item{sv}{A character vector with date(time) expressions}
}
\value{
A vector of \code{Datetime} objects with \sQuote{POSIXct} elements.
}
\description{
This function uses the Boost Date_Time library to parse
datetimes (and dates) from strings. It returns a
\code{DatetimeVector} ie a vector of \sQuote{POSIXct}
objects.
}
\details{
A numer of fixed formats are tried in succession. These include
the standard ISO format \sQuote{YYYY-MM-DD HH:MM:SS} as well as
different local variants including several forms popular in the
United States.  Two-digits years and clearly ambigous formats such
as \sQuote{03/04/05} are ignored.  In the case of parsing failure
a \code{NA} value is returned.

Fractional seconds are supported as well.  As R itself only supports
microseconds, the Boost compile-time option for nano-second resolution
has not been enabled.
}
\examples{
times <- c("2004-03-21 12:45:33.123456",
          "2004/03/21 12:45:33.123456",
          "20040321 124533.123456",
          "21.03.2004 12:45:33.123456",
          "03/21/2004 12:45:33.123456",
          "03-21-2004 12:45:33.123456",
          "2004-03-21",
          "20040321",
          "03/21/2004",
          "03-21-2004")
parsePOSIXt(times)
format(parsePOSIXt(times), tz="UTC")
}
\author{
Dirk Eddelbuettel
}

