\name{RcppBDT-package}
\alias{RcppBDT-package}
\alias{RcppBDT}
\docType{package}
\title{Bindings for Boost Date_Time}
\description{
  This package provides R with access to Boost Date_Time
  functonality by using Rcpp modules.

  Date, Local time, duration and time zone functionality is covered.
}
\details{
  Please consult the Boost documentation for (copious) details on the
  Date_Time library.
}
\author{Dirk Eddelbuettel \email{edd@debian.org}}
\references{Boost Date_Time: \url{http://www.boost.org/doc/html/date_time.html}}
\keyword{package}

